requirements.txt
flask>=3.0.0
pandas>=2.0.0
numpy>=1.25.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
openai>=1.33.0
mysql-connector-python>=9.0.0

